# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ISpinalApp": {
				"properties": {
					"name": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"iconColor": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"categoryName": {
						"type": "string"
					},
					"groupName": {
						"type": "string"
					},
					"hasViewer": {
						"type": "boolean"
					},
					"documentationLink": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"isExternalApp": {
						"type": "boolean"
					},
					"link": {
						"type": "string"
					},
					"subApps": {
						"items": {
							"$ref": "#/components/schemas/ISubApp"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"icon",
					"description",
					"tags",
					"categoryName",
					"groupName"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BOS config admin API routes",
		"version": "1.0.0",
		"description": "Welcome to the reference documentation for the Spinalcore Digital Twin REST API! </br></br>To learn about the common use cases and concept of Spinalcore REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to Spinalcore API Reference, we also provide complete documentation for BOS config adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/create_admin_app": {
			"post": {
				"operationId": "CreateAdminApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISpinalApp"
							}
						}
					}
				}
			}
		},
		"/api/v1/create_building_app": {
			"post": {
				"operationId": "CreateBuildingApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISpinalApp"
							}
						}
					}
				}
			}
		},
		"/api/v1/create_building_sub_app/{appId}": {
			"post": {
				"operationId": "CreateBuildingSubApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISubApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISubApp"
							}
						}
					}
				}
			}
		},
		"/api/v1/get_all_admin_apps": {
			"get": {
				"operationId": "GetAllAdminApps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/ISpinalApp"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_all_building_apps": {
			"get": {
				"operationId": "GetAllBuildingApps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/ISpinalApp"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_all_apps": {
			"get": {
				"operationId": "GetAllApps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"adminApps": {
													"items": {
														"$ref": "#/components/schemas/ISpinalApp"
													},
													"type": "array"
												},
												"apps": {
													"items": {
														"$ref": "#/components/schemas/ISpinalApp"
													},
													"type": "array"
												}
											},
											"required": [
												"adminApps",
												"apps"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_admin_app/{appNameOrId}": {
			"get": {
				"operationId": "GetAdminApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appNameOrId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/get_building_app/{appNaneOrId}": {
			"get": {
				"operationId": "GetBuildingApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get building app by name or id",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "app name or id",
						"in": "path",
						"name": "appNaneOrId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/get_building_sub_app/{appNameOrId}/{subAppNameOrId}": {
			"get": {
				"operationId": "GetBuildingSubApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get building sub app configuration by name or id",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appNameOrId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "subAppNameOrId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/update_admin_app/{appId}": {
			"put": {
				"operationId": "UpdateAdminApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISpinalApp"
							}
						}
					}
				}
			}
		},
		"/api/v1/update_building_app/{appId}": {
			"put": {
				"operationId": "UpdateBuildingApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISpinalApp"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISpinalApp"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}